refactor: reduced size of type

+++ structs/external-api-responses/wheater-forecast.go
@@ -3,7 +3,7 @@ package ApiResponses
 type WheaterForecast struct {
     Hourly struct {
         Time []string `json:"time"`
-        Temperature []float32 `json:"temperature_2m"`
+        Temperature []float64 `json:"temperature_2m"`
         Humidity []int8`json:"relativehumidity_2m"`
     } `json:"hourly"`
 }
